Check if stdin is a tty before reading from it
This way if the user writes

    ❯ multi-grep foo

Thinking that this will do something like search the entire project for
"foo", they'll get a message saying why it seems like the program is
hanging.

We already capture stderr in tests, so the absence of this message in
tests asserts that it doesn't infect correct usage.  But I can't figure
out a way to write a test that includes this message, because I think
the test would necessarily hang.